Split out DocSite.Console.
