+++ DocSite.sln
@@ -12,6 +12,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Jenkinsfile = Jenkinsfile
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DocSite.Console", "DocSite.Console\DocSite.Console.csproj", "{19D0FADF-EA6F-4D58-A379-279A837D62B5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -26,6 +28,10 @@ Global
 		{BDD63B8A-2F8F-4FCA-92A0-8EFF44776B35}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BDD63B8A-2F8F-4FCA-92A0-8EFF44776B35}.Release|Any CPU.ActiveCfg = Debug|Any CPU
 		{BDD63B8A-2F8F-4FCA-92A0-8EFF44776B35}.Release|Any CPU.Build.0 = Debug|Any CPU
+		{19D0FADF-EA6F-4D58-A379-279A837D62B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19D0FADF-EA6F-4D58-A379-279A837D62B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19D0FADF-EA6F-4D58-A379-279A837D62B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19D0FADF-EA6F-4D58-A379-279A837D62B5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
