+++ DocSite/Arguments.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using PTrampert.AppArgs.Attributes;
 
 namespace DocSite
 {
@@ -20,7 +19,6 @@ namespace DocSite
         /// The input file. Setting this converts the value to a full path.
         /// </summary>
         /// <value>Gets/Sets the <see cref="DocXml"/></value>
-        [Option(Name = "inputFile", ShortName = "i", Description = "The documentation xml file to process.")]
         public string DocXml
         {
             get { return _docXml; }
@@ -31,7 +29,6 @@ namespace DocSite
         /// The renderer to use.
         /// </summary>
         /// <value>Gets/Sets the <see cref="Renderer"/></value>
-        [Option(Name="renderer", ShortName = "r", Description = "The renderer to use. Defaults to Html.")]
         public RendererOptions Renderer { get; set; }
 
         private string _outDir;
@@ -40,7 +37,6 @@ namespace DocSite
         /// The directory to render the site to. This is expanded to a full path.
         /// </summary>
         /// <value>Gets/Sets the <see cref="OutputDirectory"/></value>
-        [Option(Name = "outputDirectory", ShortName = "o", Description = "Directory to render to. Defaults to the current working directory.")]
         public string OutputDirectory
         {
             get { return _outDir; }
@@ -48,17 +44,9 @@ namespace DocSite
         }
 
         /// <summary>
-        /// Flag indicating whether or not to show help.
-        /// </summary>
-        /// <value>Gets/Sets the <see cref="Help"/> flag.</value>
-        [Option(Name = "help", ShortName = "h", Description = "Show the help output.")]
-        public bool Help { get; set; }
-
-        /// <summary>
         /// Property specifying the minimum LogLevel to output. Default is Information.
         /// </summary>
         /// <value>Gets/Sets the <see cref="LogLevel"/></value>
-        [Option(Name = "logLevel", ShortName = "ll")]
         public LogLevel LogLevel { get; internal set; }
 
         /// <summary>
